third_party/WebKit/LayoutTests/media/video-buffered-unknown-duration-expected.txt

Load a video with an infinite duration. Start playback and ensure video.currentTime < video.buffered.end(0) upon first timeupdate.

EVENT(loadeddata)
EXPECTED (video.buffered.length == '1') OK
EXPECTED (video.buffered.start(0) >= '0') OK
EXPECTED (video.buffered.end(0) != 'Infinity') OK
EXPECTED (video.currentTime == '0') OK
EXPECTED (video.duration == 'Infinity') OK
EVENT(timeupdate)
EXPECTED (video.duration == 'Infinity') OK
EXPECTED (video.buffered.start(0) >= '0') OK
EXPECTED (video.buffered.end(0) < '10') OK
TEST(video.currentTime <= video.buffered.end(0)) OK
END OF TEST
